Delete unnecessary register service
